move scripts to their own folder
